user_service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Muhammad Adnan <adnasir607@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
sqlmodel = {extras = ["postgresql"], version = "^0.0.21"}
aiokafka = "^0.11.0"
protobuf = "^5.27.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["postgresql"], version = "^2.8.2"}
docker = "^7.1.0"
python-docker = "^0.2.0"
psycopg2 = "^2.9.9"
psycopg = "^3.2.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
